Move the preprocessed columns to stwo_cairo_common
