+++ stwo_cairo_prover/crates/cairo-air/Cargo.toml
@@ -22,7 +22,6 @@ stwo-cairo-common.workspace = true
 stwo-cairo-adapter.workspace = true
 rayon.workspace = true
 serde.workspace = true
-starknet-curve = "0.5.1"
 starknet-ff.workspace = true
 starknet-types-core.workspace = true
 stwo-cairo-serialize = { path = "../cairo-serialize" }
