+++ stwo_cairo_prover/crates/cairo-air/src/blake/mod.rs
@@ -1,3 +1,1 @@
-pub mod const_columns;
-pub use const_columns::BlakeSigma;
 pub mod air;
