+++ stwo_cairo_prover/crates/cairo-air/src/components/memory_address_to_id.rs
@@ -3,10 +3,10 @@ use stwo::core::channel::Channel;
 use stwo::core::fields::m31::M31;
 use stwo::core::fields::qm31::{SecureField, SECURE_EXTENSION_DEGREE};
 use stwo::core::pcs::TreeVec;
+use stwo_cairo_common::preprocessed_columns::preprocessed_trace::{PreProcessedColumn, Seq};
 use stwo_cairo_serialize::{CairoDeserialize, CairoSerialize};
 use stwo_constraint_framework::{EvalAtRow, FrameworkComponent, FrameworkEval, RelationEntry};
 
-use crate::preprocessed::{PreProcessedColumn, Seq};
 use crate::relations;
 
 /// Split the (ID , Multiplicity) columns to shorter chunks. This is done to improve the performance
