+++ stwo_cairo_prover/crates/cairo-air/src/components/memory_id_to_big.rs
@@ -7,13 +7,13 @@ use stwo::core::fields::qm31::{SecureField, SECURE_EXTENSION_DEGREE};
 use stwo::core::pcs::TreeVec;
 use stwo_cairo_adapter::memory::LARGE_MEMORY_VALUE_ID_BASE;
 use stwo_cairo_common::memory::{N_M31_IN_FELT252, N_M31_IN_SMALL_FELT252};
+use stwo_cairo_common::preprocessed_columns::preprocessed_trace::{PreProcessedColumn, Seq};
 use stwo_cairo_serialize::{CairoDeserialize, CairoSerialize};
 use stwo_constraint_framework::{
     relation, EvalAtRow, FrameworkComponent, FrameworkEval, RelationEntry, TraceLocationAllocator,
 };
 
 use super::prelude::RelationUse;
-use crate::preprocessed::{PreProcessedColumn, Seq};
 use crate::relations;
 
 // TODO(AlonH): Make memory size configurable.
