+++ stwo_cairo_prover/crates/cairo-air/src/pedersen/mod.rs
@@ -1,3 +1,1 @@
 pub mod air;
-pub mod const_columns;
-pub mod utils;
