+++ stwo_cairo_prover/crates/common/src/lib.rs
@@ -1,4 +1,4 @@
 #![feature(portable_simd)]
 pub mod memory;
-pub mod preprocessed_consts;
+pub mod preprocessed_columns;
 pub mod prover_types;
