+++ stwo_cairo_prover/crates/common/src/preprocessed_columns/preprocessed_utils.rs
@@ -1,4 +1,8 @@
+use std::simd::Simd;
+
 use stwo::core::fields::m31::M31;
+
+use crate::prover_types::simd::N_LANES;
 
 // Pads all rows below <padding_offset> with the first row. Uses the <get_m31> function to get the
 // value in a given row and column.
@@ -12,3 +16,6 @@ where
         .map(|i| get_m31(i, col))
         .collect()
 }
+
+pub const SIMD_ENUMERATION_0: Simd<u32, N_LANES> =
+    Simd::from_array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]);
