--- stwo_cairo_prover/crates/common/src/preprocessed_consts/mod.rs
@@ -1,3 +1,0 @@
-pub mod blake;
-pub mod pedersen;
-pub mod poseidon;
