+++ stwo_cairo_prover/crates/prover/src/prover.rs
@@ -253,9 +253,9 @@ pub fn create_and_serialize_proof(
 
 #[cfg(test)]
 pub mod tests {
-    use cairo_air::preprocessed::testing_preprocessed_tree;
     use dev_utils::utils::get_compiled_cairo_program_path;
     use stwo_cairo_adapter::utils::{run_program_and_adapter, ProgramType};
+    use stwo_cairo_common::preprocessed_columns::preprocessed_trace::testing_preprocessed_tree;
 
     use crate::debug_tools::assert_constraints::assert_cairo_constraints;
     #[test]
@@ -353,13 +353,13 @@ pub mod tests {
         use std::io::Write;
         use std::process::Command;
 
-        use cairo_air::preprocessed::PreProcessedTrace;
         use cairo_air::verifier::verify_cairo;
         use dev_utils::utils::{get_compiled_cairo_program_path, get_proof_file_path};
         use itertools::Itertools;
         use stwo::core::fri::FriConfig;
         use stwo::core::pcs::PcsConfig;
         use stwo::core::vcs::blake2_merkle::Blake2sMerkleChannel;
+        use stwo_cairo_common::preprocessed_columns::preprocessed_trace::PreProcessedTrace;
         use stwo_cairo_serialize::CairoSerialize;
         use tempfile::NamedTempFile;
         use test_log::test;
@@ -493,6 +493,7 @@ pub mod tests {
         /// These tests' inputs were generated using cairo-vm with 50 instances of each builtin.
         pub mod builtin_tests {
             use stwo_cairo_adapter::utils::{run_program_and_adapter, ProgramType};
+            use stwo_cairo_common::preprocessed_columns::preprocessed_trace::testing_preprocessed_tree;
             use test_log::test;
 
             use super::*;
