+++ stwo_cairo_prover/crates/prover/src/witness/fast_deduction/blake.rs
@@ -2,7 +2,7 @@ use std::array::from_fn;
 use std::simd::u32x16;
 
 use stwo_cairo_adapter::memory::Memory;
-use stwo_cairo_common::preprocessed_consts::blake::{BLAKE_SIGMA, N_BLAKE_SIGMA_COLS};
+use stwo_cairo_common::preprocessed_columns::blake::{BLAKE_SIGMA, N_BLAKE_SIGMA_COLS};
 use stwo_cairo_common::prover_types::simd::{PackedUInt32, N_LANES};
 
 use crate::stwo::prover::backend::simd::m31::PackedM31;
