+++ stwo_cairo_prover/crates/prover/src/witness/prelude.rs
@@ -1,9 +1,6 @@
 pub use std::iter::zip;
 pub use std::simd::Simd;
 
-pub use cairo_air::pedersen::const_columns::PedersenPoints;
-pub use cairo_air::poseidon::const_columns::PoseidonRoundKeys;
-pub use cairo_air::preprocessed::*;
 pub(crate) use cairo_air::relations;
 pub use num_traits::{One, Zero};
 pub use rayon::prelude::*;
@@ -20,6 +17,10 @@ pub use stwo::prover::poly::circle::CircleEvaluation;
 pub use stwo::prover::poly::BitReversedOrder;
 pub use stwo_air_utils::trace::component_trace::ComponentTrace;
 pub use stwo_air_utils_derive::{IterMut, ParIterMut, Uninitialized};
+pub use stwo_cairo_common::preprocessed_columns::bitwise_xor::BitwiseXor;
+pub use stwo_cairo_common::preprocessed_columns::pedersen::PedersenPoints;
+pub use stwo_cairo_common::preprocessed_columns::poseidon::PoseidonRoundKeys;
+pub use stwo_cairo_common::preprocessed_columns::preprocessed_trace::Seq;
 pub use stwo_cairo_common::prover_types::cpu::*;
 pub use stwo_cairo_common::prover_types::simd::*;
 pub use stwo_constraint_framework::{LogupTraceGenerator, Relation};
